requirements.txt
streamlit
langchain
openai
python-dotenv
langchain-community
PyMySQL
groq
langchain-groq